my_prj1/tg-bot-AI-0.1/requirements.txt
3==0.0.0
aiofiles==23.2.1
aiogram==3.16.0
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
asyncpg==0.30.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
environs==11.0.0
et-xmlfile==1.1.0
executing==2.0.1
fake-useragent==1.5.1
fonttools==4.50.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lexicon==2.0.1
lxml==5.2.1
magic-filter==1.0.12
marshmallow==3.23.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
parso==0.8.3
patsy==0.5.6
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.10
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.2
pyTelegramBotAPI==4.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
redis==5.2.1
requests==2.31.0
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.38
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
